third_party/blink/web_tests/fast/forms/state-save-of-detached-control-expected.txt

Test to NOT save state of a detached form control.

PASS document.getElementById("input2").value is ""
PASS successfullyParsed is true

TEST COMPLETE
